refactor(frontend): added close icon and used icon component bb-106
