+++ apps/frontend/src/libs/components/icon/libs/maps/icon-name-to-svg.map.ts
@@ -1,3 +1,5 @@
+import Close from "~/assets/img/close.svg?react";
+import Menu from "~/assets/img/menu.svg?react";
 import WheelOfBalanceActive from "~/assets/img/wheel-of-balance-active.svg?react";
 import WheelOfBalanceInactive from "~/assets/img/wheel-of-balance-inactive.svg?react";
 import { type IconName } from "~/libs/types/types.js";
@@ -6,6 +8,8 @@ const iconNameToSvg: Record<
 	IconName,
 	React.FC<React.SVGProps<SVGSVGElement>>
 > = {
+	close: Close,
+	menu: Menu,
 	wheelActive: WheelOfBalanceActive,
 	wheelInactive: WheelOfBalanceInactive,
 };
