+++ apps/frontend/src/libs/components/sidebar/sidebar.tsx
@@ -1,3 +1,4 @@
+import { Icon } from "~/libs/components/components.js";
 import { SIDEBAR_ITEMS } from "~/libs/constants/constants.js";
 import { getValidClassNames } from "~/libs/helpers/helpers.js";
 import { useLocation } from "~/libs/hooks/hooks.js";
@@ -7,9 +8,13 @@ import styles from "./styles.module.css";
 
 type Properties = {
 	isSidebarOpen: boolean;
+	toggleSidebar: () => void;
 };
 
-const Sidebar: React.FC<Properties> = ({ isSidebarOpen }: Properties) => {
+const Sidebar: React.FC<Properties> = ({
+	isSidebarOpen,
+	toggleSidebar,
+}: Properties) => {
 	const { pathname } = useLocation();
 
 	return (
@@ -19,9 +24,14 @@ const Sidebar: React.FC<Properties> = ({ isSidebarOpen }: Properties) => {
 				isSidebarOpen && styles["open"],
 			)}
 		>
-			<label className={styles["close-btn"]} htmlFor="sidebar-toggler">
+			{/* <label className={styles["close-btn"]} htmlFor="sidebar-toggler">
 				x
-			</label>
+			</label> */}
+
+			<button className={styles["close-btn"]} onClick={toggleSidebar}>
+				<Icon name="close" />
+			</button>
+
 			<div className={styles["logo-container"]}>Logo</div>
 			<div className={styles["buttons-container"]}>
 				{SIDEBAR_ITEMS.map(({ href, icon, label }) => {
