+++ apps/frontend/src/libs/types/icon-name.type.ts
@@ -1,3 +1,3 @@
-type IconName = "wheelActive" | "wheelInactive";
+type IconName = "close" | "menu" | "wheelActive" | "wheelInactive";
 
 export { type IconName };
